add params to apps layout loader, add conditionals to favorites and application Model call in controller, and add sort to get Applications model
